# openAPI.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Dataloaders API",
        "description": "API documentation for Dataloaders application",
        "contact": {
            "name": "Dataloaders Team",
            "email": "support@dataloaders.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/dataloader/job/{jobId}": {
            "get": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Get job config",
                "description": "Get job configuration by ID",
                "operationId": "get_2",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job config found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job config not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Update job config",
                "description": "Update job configuration",
                "operationId": "update_2",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Job config updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job config not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Delete job config",
                "description": "Delete job configuration",
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job config deleted successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job config not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/job/publish": {
            "post": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Publish job config",
                "description": "Publish job configuration",
                "operationId": "publish",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Job config published successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/job/draft": {
            "post": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Save draft",
                "description": "Save job configuration as draft",
                "operationId": "saveDraft",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Draft saved successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/job/deploy/{jobId}": {
            "post": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Deploy job",
                "description": "Deploy a job configuration",
                "operationId": "deploy",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deployed successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/resource/{resourceId}/status": {
            "patch": {
                "tags": [
                    "Resources"
                ],
                "summary": "Update resource status",
                "description": "Activate or deactivate a resource",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "resourceId",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Active status",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/job/{jobId}/status": {
            "patch": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Update job config status",
                "description": "Update job configuration status",
                "operationId": "updateStatus_2",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Active status",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/configFields/{resourceTypeId}/status": {
            "patch": {
                "tags": [
                    "Field Definitions"
                ],
                "summary": "Update field definition status",
                "description": "Activate or deactivate a field definition",
                "operationId": "updateStatus_3",
                "parameters": [
                    {
                        "name": "resourceTypeId",
                        "in": "path",
                        "description": "Field definition ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Active status",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Field definition not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/job/writer-configs/{resourceType}": {
            "get": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Get writer config details",
                "description": "Get writer configuration details by resource type",
                "operationId": "getWriterConfigDetails",
                "parameters": [
                    {
                        "name": "resourceType",
                        "in": "path",
                        "description": "Resource type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Writer config details retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/job/reader-configs/{resourceType}": {
            "get": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Get reader config details",
                "description": "Get reader configuration details by resource type",
                "operationId": "getReaderConfigDetails",
                "parameters": [
                    {
                        "name": "resourceType",
                        "in": "path",
                        "description": "Resource type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reader config details retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/job/parent/{jobId}": {
            "get": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Get job configs by parent",
                "description": "Get job configurations by parent job ID",
                "operationId": "getByParent",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Parent Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job configs retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/job/dataflow/{itemId}": {
            "get": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Get job configs by item",
                "description": "Get all job configurations for an item",
                "operationId": "getAllByItemId",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Item ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job configs retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/deploy/job/{jobId}": {
            "get": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Get deployments by job ID",
                "description": "Get all deployments for a job",
                "operationId": "getDeploysByJobId",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deployments retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Deploy"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Response": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "causedBy": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDefinition"
                        }
                    }
                }
            },
...
            "Deploy": {
                "type": "object",
                "properties": {
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "updatedBy": {
                        "type": "string"
                    },
                    "deployId": {
                        "type": "string"
                    },
                    "parentJobId": {
                        "type": "string"
                    },
                    "jobId": {
                        "type": "string"
                    },
                    "deployName": {
                        "type": "string"
                    },
                    "manualRun": {
                        "type": "boolean"
                    },
                    "schedule": {
                        "type": "boolean"
                    },
                    "scheduleExpression": {
                        "type": "string"
                    },
                    "schedulerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "schedulerName": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}